move request feed to files
